third_party/blink/web_tests/external/wpt/html/semantics/popups/popup-events.tentative.html

<!DOCTYPE html>
<meta charset="utf-8" />
<title>Popup events</title>
<link rel="author" href="mailto:masonf@chromium.org">
<link rel=help href="https://open-ui.org/components/popup.research.explainer">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/popup-utils.js"></script>

<div popup>Popup</div>

<script>
window.onload = () => {
  promise_test(async t => {
    const popup = document.querySelector('[popup]');
    let showCount = 0;
    let hideCount = 0;
    assert_false(popup.matches(':open'));
    const controller = new AbortController();
    const signal = controller.signal;
    t.add_cleanup(() => controller.abort());
    document.addEventListener('show',() => ++showCount, {signal});
    document.addEventListener('hide',() => ++hideCount, {signal});
    assert_equals(0,showCount);
    assert_equals(0,hideCount);
    popup.showPopUp();
    assert_true(popup.matches(':open'));
    assert_equals(1,showCount);
    assert_equals(0,hideCount);
    await waitForRender();
    assert_true(popup.matches(':open'));
    popup.hidePopUp();
    assert_false(popup.matches(':open'));
    assert_equals(1,showCount);
    assert_equals(1,hideCount);
    await waitForRender();
    // No additional events after animation frame
    assert_false(popup.matches(':open'));
    assert_equals(1,showCount);
    assert_equals(1,hideCount);
  }, 'Show and hide events get properly dispatched for popups');

  promise_test(async t => {
    const popUp = document.querySelector('[popup]');
    const controller = new AbortController();
    const signal = controller.signal;
    t.add_cleanup(() => controller.abort());
    let cancel = true;
    popUp.addEventListener('show',(e) => {
      if (cancel)
        e.preventDefault();
    }, {signal});
    assert_false(popUp.matches(':open'));
    popUp.showPopUp();
    assert_false(popUp.matches(':open'),'The "show" event should be cancelable');
    cancel = false;
    popUp.showPopUp();
    assert_true(popUp.matches(':open'));
    popUp.hidePopUp();
    assert_false(popUp.matches(':open'));
  }, 'Show event is cancelable');
};
</script>
